+++ uncertainties/core/core.py
@@ -19,15 +19,6 @@ import math
 from math import sqrt, log, isnan, isinf  # Optimization: no attribute look-up
 import re
 import sys
-if sys.version_info < (3,):
-     from past.builtins import basestring
-else:
-     # Avoid importing from past in Python 3 since it utilizes the builtin
-     # 'imp' module, which is deprecated as of Python 3.4, see
-     # https://docs.python.org/3/library/imp.html. The 2to3 tool replaces
-     # basestring with str, so that's what we effectively do here as well:
-     basestring = str
-
 try:
     from math import isinfinite  # !! Python 3.2+
 except ImportError:
@@ -48,189 +39,21 @@ if sys.version_info < (3,):  # !! Could be removed when moving to Python 3 only
 else:
     from inspect import getfullargspec as getargspec
 
-###############################################################################
-
-def set_doc(doc_string):
-    """
-    Decorator function that sets the docstring to the given text.
-
-    It is useful for functions whose docstring is calculated
-    (including string substitutions).
-    """
-    def set_doc_string(func):
-        func.__doc__ = doc_string
-        return func
-    return set_doc_string
-
-# Some types known to not depend on Variable objects are put in
-# CONSTANT_TYPES.  The most common types can be put in front, as this
-# may slightly improve the execution speed.
-FLOAT_LIKE_TYPES = (numbers.Number,)
-CONSTANT_TYPES = FLOAT_LIKE_TYPES+(complex,)
-
-###############################################################################
-
-# Utility for issuing deprecation warnings
-
-def deprecation(message):
-    '''
-    Warn the user with the given message, by issuing a
-    DeprecationWarning.
-    '''
-
-    # stacklevel = 3 points to the original user call (not to the
-    # function from this module that called deprecation()).
-    # DeprecationWarning is ignored by default: not used.
-
-    warnings.warn('Obsolete: %s Code can be automatically updated with'
-                  ' python -m uncertainties.1to2 -w ProgramDirectory.'
-                  % message, stacklevel=3)
-
-###############################################################################
-
-## Definitions that depend on the availability of NumPy:
-
+from . compat import (deprecation, FLOAT_LIKE_TYPES, CONSTANT_TYPES, CallableStdDev,
+)
+from . util import (set_doc, covariance_matrix, partial_derivative,
+NumericalDerivatives, IndexableIter, nominal_value, std_dev, basestring,
+)
+from . formatting import (first_digit, PDG_precision, format_num, signif_dgt_to_limit
+)
+from . parsing import (str_to_number_with_uncert,)
 
 try:
     import numpy
 except ImportError:
     pass
 else:
-
-    # NumPy numbers do not depend on Variable objects:
-    FLOAT_LIKE_TYPES += (numpy.generic,)
-    CONSTANT_TYPES += FLOAT_LIKE_TYPES[-1:]
-
-    # Entering variables as a block of correlated values.  Only available
-    # if NumPy is installed.
-
-    #! It would be possible to dispense with NumPy, but a routine should be
-    # written for obtaining the eigenvectors of a symmetric matrix.  See
-    # for instance Numerical Recipes: (1) reduction to tri-diagonal
-    # [Givens or Householder]; (2) QR / QL decomposition.
-
-    def correlated_values(nom_values, covariance_mat, tags=None):
-        """
-        Return numbers with uncertainties (AffineScalarFunc objects)
-        that correctly reproduce the given covariance matrix, and have
-        the given (float) values as their nominal value.
-
-        The correlated_values_norm() function returns the same result,
-        but takes a correlation matrix instead of a covariance matrix.
-
-        The list of values and the covariance matrix must have the
-        same length, and the matrix must be a square (symmetric) one.
-
-        The numbers with uncertainties returned depend on newly
-        created, independent variables (Variable objects).
-
-        nom_values -- sequence with the nominal (real) values of the
-        numbers with uncertainties to be returned.
-
-        covariance_mat -- full covariance matrix of the returned numbers with
-        uncertainties. For example, the first element of this matrix is the
-        variance of the first number with uncertainty. This matrix must be a
-        NumPy array-like (list of lists, NumPy array, etc.). 
-
-        tags -- if 'tags' is not None, it must list the tag of each new
-        independent variable.
-        """
-
-        # !!! It would in principle be possible to handle 0 variance
-        # variables by first selecting the sub-matrix that does not contain
-        # such variables (with the help of numpy.ix_()), and creating 
-        # them separately.
-        
-        std_devs = numpy.sqrt(numpy.diag(covariance_mat))
-
-        # For numerical stability reasons, we go through the correlation
-        # matrix, because it is insensitive to any change of scale in the
-        # quantities returned. However, care must be taken with 0 variance
-        # variables: calculating the correlation matrix cannot be simply done
-        # by dividing by standard deviations. We thus use specific
-        # normalization values, with no null value:
-        norm_vector = std_devs.copy()
-        norm_vector[norm_vector==0] = 1
-
-        return correlated_values_norm(
-            # !! The following zip() is a bit suboptimal: correlated_values()
-            # separates back the nominal values and the standard deviations:
-            list(zip(nom_values, std_devs)),
-            covariance_mat/norm_vector/norm_vector[:,numpy.newaxis],
-            tags)
-
-    __all__.append('correlated_values')
-
-    def correlated_values_norm(values_with_std_dev, correlation_mat,
-                               tags=None):
-        '''
-        Return correlated values like correlated_values(), but takes
-        instead as input:
-
-        - nominal (float) values along with their standard deviation, and
-        - a correlation matrix (i.e. a normalized covariance matrix).
-
-        values_with_std_dev -- sequence of (nominal value, standard
-        deviation) pairs. The returned, correlated values have these
-        nominal values and standard deviations.
-
-        correlation_mat -- correlation matrix between the given values, except
-        that any value with a 0 standard deviation must have its correlations
-        set to 0, with a diagonal element set to an arbitrary value (something
-        close to 0-1 is recommended, for a better numerical precision).  When
-        no value has a 0 variance, this is the covariance matrix normalized by
-        standard deviations, and thus a symmetric matrix with ones on its
-        diagonal.  This matrix must be an NumPy array-like (list of lists,
-        NumPy array, etc.).
-
-        tags -- like for correlated_values().
-        '''
-
-        # If no tags were given, we prepare tags for the newly created
-        # variables:
-        if tags is None:
-            tags = (None,) * len(values_with_std_dev)
-
-        (nominal_values, std_devs) = numpy.transpose(values_with_std_dev)
-
-        # We diagonalize the correlation matrix instead of the
-        # covariance matrix, because this is generally more stable
-        # numerically. In fact, the covariance matrix can have
-        # coefficients with arbitrary values, through changes of units
-        # of its input variables. This creates numerical instabilities.
-        #
-        # The covariance matrix is diagonalized in order to define
-        # the independent variables that model the given values:
-        (variances, transform) = numpy.linalg.eigh(correlation_mat)
-
-        # Numerical errors might make some variances negative: we set
-        # them to zero:
-        variances[variances < 0] = 0.
-
-        # Creation of new, independent variables:
-
-        # We use the fact that the eigenvectors in 'transform' are
-        # special: 'transform' is unitary: its inverse is its transpose:
-
-        variables = tuple(
-            # The variables represent "pure" uncertainties:
-            Variable(0, sqrt(variance), tag)
-            for (variance, tag) in zip(variances, tags))
-
-        # The coordinates of each new uncertainty as a function of the
-        # new variables must include the variable scale (standard deviation):
-        transform *= std_devs[:, numpy.newaxis] 
-        
-        # Representation of the initial correlated values:
-        values_funcs = tuple(
-            AffineScalarFunc(
-                value,
-                LinearCombination(dict(zip(variables, coords))))
-            for (coords, value) in zip(transform, nominal_values))
-
-        return values_funcs
-
-    __all__.append('correlated_values_norm')
+    from . util import correlation_matrix, correlated_values, correlated_values_norm
 
 ###############################################################################
 
@@ -262,149 +85,6 @@ def to_affine_scalar(x):
     raise NotUpcast("%s cannot be converted to a number with"
                     " uncertainty" % type(x))
 
-# Step constant for numerical derivatives in
-# partial_derivative(). Value chosen to as to get better numerical
-# results:
-STEP_SIZE = sqrt(sys.float_info.epsilon)
-
-# !! It would be possible to split the partial derivative calculation
-# into two functions: one for positional arguments (case of integer
-# arg_ref) and one for keyword arguments (case of string
-# arg_ref). However, this would either duplicate the code for the
-# numerical differentiation, or require a call, which is probably more
-# expensive in time than the tests done here.
-def partial_derivative(f, arg_ref):
-    """
-    Return a function that numerically calculates the partial
-    derivative of function f with respect to its argument arg_ref.
-
-    arg_ref -- describes which variable to use for the
-    differentiation. If f is called with f(*args, **kwargs) arguments,
-    an integer represents the index of an argument in args, and a
-    string represents the name of an argument in kwargs.
-    """
-
-    # Which set of function parameter contains the variable to be
-    # changed? the positional or the optional keyword arguments?
-    change_kwargs = isinstance(arg_ref, basestring)
-
-    def partial_derivative_of_f(*args, **kwargs):
-        """
-        Partial derivative, calculated with the (-epsilon, +epsilon)
-        method, which is more precise than the (0, +epsilon) method.
-        """
-
-        # args_with_var contains the arguments (either args or kwargs)
-        # that contain the variable that must be shifted, as a mutable
-        # object (because the variable contents will be modified):
-
-        # The values in args need to be modified, for the
-        # differentiation: it is converted to a list:
-        if change_kwargs:
-            args_with_var = kwargs
-        else:
-            args_with_var = list(args)
-
-        # The step is relative to the parameter being varied, so that
-        # shifting it does not suffer from finite precision limitations:
-        step = STEP_SIZE*abs(args_with_var[arg_ref])
-        if not step:
-            # Arbitrary, but "small" with respect to 1:
-            step = STEP_SIZE
-
-        args_with_var[arg_ref] += step
-
-        if change_kwargs:
-            shifted_f_plus = f(*args, **args_with_var)
-        else:
-            shifted_f_plus = f(*args_with_var, **kwargs)
-
-        args_with_var[arg_ref] -= 2*step  # Optimization: only 1 list copy
-
-        if change_kwargs:
-            shifted_f_minus = f(*args, **args_with_var)
-        else:
-            shifted_f_minus = f(*args_with_var, **kwargs)
-
-        return (shifted_f_plus - shifted_f_minus)/2/step
-
-    return partial_derivative_of_f
-
-class NumericalDerivatives(object):
-    """
-    Convenient access to the partial derivatives of a function,
-    calculated numerically.
-    """
-    # This is not a list because the number of arguments of the
-    # function is not known in advance, in general.
-
-    def __init__(self, function):
-        """
-        'function' is the function whose derivatives can be computed.
-        """
-        self._function = function
-
-    def __getitem__(self, n):
-        """
-        Return the n-th numerical derivative of the function.
-        """
-        return partial_derivative(self._function, n)
-
-class IndexableIter(object):
-    '''
-    Iterable whose values can also be accessed through indexing.
-
-    The input iterable values are cached.
-
-    Some attributes:
-
-    iterable -- iterable used for returning the elements one by one.
-
-    returned_elements -- list with the elements directly accessible.
-    through indexing. Additional elements are obtained from self.iterable.
-
-    none_converter -- function that takes an index and returns the
-    value to be returned when None is obtained form the iterable
-    (instead of None).
-    '''
-
-    def __init__(self, iterable, none_converter=lambda index: None):
-        '''
-        iterable -- iterable whose values will be returned.
-
-        none_converter -- function applied to None returned
-        values. The value that replaces None is none_converter(index),
-        where index is the index of the element.
-        '''
-        self.iterable = iterable
-        self.returned_elements = []
-        self.none_converter = none_converter
-
-    def __getitem__(self, index):
-
-        returned_elements = self.returned_elements
-
-        try:
-
-            return returned_elements[index]
-
-        except IndexError:  # Element not yet cached
-
-            for pos in range(len(returned_elements), index+1):
-
-                value = next(self.iterable)
-
-                if value is None:
-                    value = self.none_converter(pos)
-
-                returned_elements.append(value)
-
-            return returned_elements[index]
-
-    def __str__(self):
-        return '<%s: [%s...]>' % (
-            self.__class__.__name__,
-            ', '.join(map(str, self.returned_elements)))
 
 def wrap(f, derivatives_args=[], derivatives_kwargs={}):
     """
@@ -843,629 +523,7 @@ def le_on_aff_funcs(self, y_with_uncert):
     return (lt_on_aff_funcs(self, y_with_uncert)
             or eq_on_aff_funcs(self, y_with_uncert))
 
-########################################
 
-def first_digit(value):
-    '''
-    Return the first digit position of the given value, as an integer.
-
-    0 is the digit just before the decimal point. Digits to the right
-    of the decimal point have a negative position.
-
-    Return 0 for a null value.
-    '''
-    try:
-        return int(math.floor(math.log10(abs(value))))
-    except ValueError:  # Case of value == 0
-        return 0
-
-def PDG_precision(std_dev):
-    '''
-    Return the number of significant digits to be used for the given
-    standard deviation, according to the rounding rules of the
-    Particle Data Group (2010)
-    (http://pdg.lbl.gov/2010/reviews/rpp2010-rev-rpp-intro.pdf).
-
-    Also returns the effective standard deviation to be used for
-    display.
-    '''
-
-    exponent = first_digit(std_dev)
-
-    # The first three digits are what matters: we get them as an
-    # integer number in [100; 999).
-    #
-    # In order to prevent underflow or overflow when calculating
-    # 10**exponent, the exponent is slightly modified first and a
-    # factor to be applied after "removing" the new exponent is
-    # defined.
-    #
-    # Furthermore, 10**(-exponent) is not used because the exponent
-    # range for very small and very big floats is generally different.
-    if exponent >= 0:
-        # The -2 here means "take two additional digits":
-        (exponent, factor) = (exponent-2, 1)
-    else:
-        (exponent, factor) = (exponent+1, 1000)
-    digits = int(std_dev/10.**exponent*factor)  # int rounds towards zero
-
-    # Rules:
-    if digits <= 354:
-        return (2, std_dev)
-    elif digits <= 949:
-        return (1, std_dev)
-    else:
-        # The parentheses matter, for very small or very large
-        # std_dev:
-        return (2, 10.**exponent*(1000/factor))
-
-# Definition of a basic (format specification only, no full-feature
-# format string) formatting function that works whatever the version
-# of Python. This function exists so that the more capable format() is
-# used instead of the % formatting operator, if available:
-robust_format = format
-
-class CallableStdDev(float):
-    '''
-    Class for standard deviation results, which used to be
-    callable. Provided for compatibility with old code. Issues an
-    obsolescence warning upon call.
-    '''
-
-    # This class is a float. It must be set to the standard deviation
-    # upon construction.
-
-    def __call__ (self):
-        deprecation('the std_dev attribute should not be called'
-                    ' anymore: use .std_dev instead of .std_dev().')
-        return self
-
-# Exponent letter: the keys are the possible main_fmt_type values of
-# format_num():
-EXP_LETTERS = {'f': 'e', 'F': 'E'}
-
-def robust_align(orig_str, fill_char, align_option, width):
-    '''
-    Aligns the given string with the given fill character.
-
-    orig_str -- string to be aligned (str or unicode object).
-
-    fill_char -- if empty, space is used.
-
-    align_option -- as accepted by format().
-
-    wdith -- string that contains the width.
-    '''
-
-    # print "ALIGNING", repr(orig_str), "WITH", fill_char+align_option,
-    # print "WIDTH", width
-
-    return format(orig_str, fill_char+align_option+width)
-
-# Maps some Unicode code points ("-", "+", and digits) to their
-# superscript version:
-TO_SUPERSCRIPT = {
-    0x2b: u'⁺',
-    0x2d: u'⁻',
-    0x30: u'⁰',
-    0x31: u'¹',
-    0x32: u'²',
-    0x33: u'³',
-    0x34: u'⁴',
-    0x35: u'⁵',
-    0x36: u'⁶',
-    0x37: u'⁷',
-    0x38: u'⁸',
-    0x39: u'⁹'
-    }
-
-# Inverted TO_SUPERSCRIPT table, for use with unicode.translate():
-#
-#! Python 2.7+ can use a dictionary comprehension instead:
-FROM_SUPERSCRIPT = {
-    ord(sup): normal for (normal, sup) in TO_SUPERSCRIPT.items()}
-
-def to_superscript(value):
-    '''
-    Return a (Unicode) string with the given value as superscript characters.
-
-    The value is formatted with the %d %-operator format.
-
-    value -- integer.
-    '''
-
-    return (u'%d' % value).translate(TO_SUPERSCRIPT)
-
-def nrmlze_superscript(number_str):
-    '''
-    Return a string with superscript digits transformed into regular digits.
-
-    Non-superscript digits are not changed before the conversion. Thus, the
-    string can also contain regular digits.
-
-    ValueError is raised if the conversion cannot be done.
-
-    number_str -- string to be converted (of type str, but also possibly, for 
-    Python 2, unicode, which allows this string to contain superscript digits).
-    '''
-    # !! Python 3 doesn't need this str(), which is only here for giving the
-    # .translate() method to str objects in Python 2 (this str() comes
-    # from the builtins module of the future package and is therefore
-    # a subclass of unicode, in Python 2):
-    return int(str(number_str).translate(FROM_SUPERSCRIPT))
-
-PM_SYMBOLS = {'pretty-print': u'±', 'latex': r' \pm ', 'default': '+/-'}
-
-# Multiplication symbol for pretty printing (so that pretty printing can
-# be customized):
-MULT_SYMBOLS = {'pretty-print': u'×', 'latex': r'\times'}
-
-# Function that transforms a numerical exponent produced by format_num() into
-# the corresponding string notation (for non-default modes):
-EXP_PRINT = {
-    'pretty-print': lambda common_exp: u'%s10%s' % (
-        MULT_SYMBOLS['pretty-print'], to_superscript(common_exp)),
-    'latex': lambda common_exp: r' %s 10^{%d}' % (
-        MULT_SYMBOLS['latex'], common_exp)}
-
-# Symbols used for grouping (typically between parentheses) in format_num():
-GROUP_SYMBOLS = {
-    'pretty-print': ('(', ')'),
-    # Because of possibly exponents inside the parentheses (case of a
-    # specified field width), it is better to use auto-adjusting
-    # parentheses. This has the side effect of making the part between
-    # the parentheses non-breakable (the text inside parentheses in a
-    # LaTeX math expression $...$ can be broken).
-    'latex': (r'\left(', r'\right)'),
-    'default': ('(', ')')  # Basic text mode
-    }
-
-def format_num(nom_val_main, error_main, common_exp,
-               fmt_parts, prec, main_pres_type, options):
-    u'''
-    Return a formatted number with uncertainty.
-
-    Null errors (error_main) are displayed as the integer 0, with
-    no decimal point.
-
-    The formatting can be customized globally through the PM_SYMBOLS,
-    MULT_SYMBOLS, GROUP_SYMBOLS and EXP_PRINT dictionaries, which contain
-    respectively the symbol for ±, for multiplication, for parentheses, and a
-    function that maps an exponent to something like "×10²" (using
-    MULT_SYMBOLS).
-
-    Each of these dictionary has (at least) a 'pretty-print' and a 'latex' key,
-    that define the symbols to be used for these two output formats (the
-    PM_SYMBOLS and GROUP_SYMBOLS also have a 'default' key for the default
-    output format). For example, the defaults for the 'pretty-print' format
-    are:
-
-    - PM_SYMBOLS['pretty-print'] = '±'
-    - MULT_SYMBOLS['pretty-print'] = '×'
-    - GROUP_SYMBOLS['pretty-print'] = ( '(', ')' )
-    - EXP_PRINT['pretty-print']: see the source code.
-
-    Arguments:
-
-    nom_val_main, error_main -- nominal value and error, before using
-    common_exp (e.g., "1.23e2" would have a main value of 1.23;
-    similarly, "12.3+/-0.01" would have a main value of 12.3).
-
-    common_exp -- common exponent to use. If None, no common exponent
-    is used.
-
-    fmt_parts -- mapping that contains at least the following parts of
-    the format specification: fill, align, sign, zero, width, comma,
-    type; the value are strings. These format specification parts are
-    handled. The width is applied to each value, or, if the shorthand
-    notation is used, globally. If the error is special (zero, NaN, inf),
-    the parts are applied as much as possible to the nominal value.
-
-    prec -- precision to use with the main_pres_type format type
-    (see below).
-
-    main_pres_type -- format presentation type, either "f" or
-    "F". This defines how the mantissas, exponents and NaN/inf values are
-    represented (in the same way as for float). None, the empty
-    string, or "%" are not accepted.
-
-    options -- options (as an object that support membership testing, like for
-    instance a string). "P" is for pretty-printing ("±" between the nominal
-    value and the error, superscript exponents, etc.). "L" is for a LaTeX
-    output. "S" is for the shorthand notation 1.23(1). "p" is for making sure
-    that the …±… part is surrounded by parentheses.  "%" adds a final percent
-    sign, and parentheses if the shorthand notation is not used. Options can
-    be combined. The P option has priority over the L option (if both are
-    given). For details, see the documentation for
-    AffineScalarFunction.__format__().
-    '''
-
-    # print (nom_val_main, error_main, common_exp,
-    #        fmt_parts, prec, main_pres_type, options)
-
-    # If a decimal point were always present in zero rounded errors
-    # that are not zero, the formatting would be difficult, in general
-    # (because the formatting options are very general): an example
-    # is'{:04.0f}'.format(0.1), which gives "0000" and would have to
-    # give "000.". Another example is '{:<4.0f}'.format(0.1), which
-    # gives "0 " but should give "0.  ". This is cumbersome to
-    # implement in the general case, because no format prints "0."
-    # for 0. Furthermore, using the .0f format already brings the same
-    # kind of difficulty: non-zero numbers can appear as the exact
-    # integer zero, after rounding. The problem is not larger, for
-    # numbers with an error.
-    #
-    # That said, it is good to indicate null errors explicitly when
-    # possible: printing 3.1±0 with the default format prints 3.1+/-0,
-    # which shows that the uncertainty is exactly zero.
-
-    # The suffix of the result is calculated first because it is
-    # useful for the width handling of the shorthand notation.
-
-    # Printing type for parts of the result (exponent, parentheses),
-    # taking into account the priority of the pretty-print mode over
-    # the LaTeX mode. This setting does not apply to everything: for
-    # example, NaN is formatted as \mathrm{nan} (or NAN) if the LaTeX
-    # mode is required.
-    if 'P' in options:
-        print_type = 'pretty-print'
-    elif 'L' in options:
-        print_type = 'latex'
-    else:
-        print_type = 'default'
-
-    # Exponent part:
-    if common_exp is None:
-        exp_str = ''
-    elif print_type == 'default':
-        # Case of e or E. The same convention as Python 2.7
-        # to 3.3 is used for the display of the exponent:
-        exp_str = EXP_LETTERS[main_pres_type]+'%+03d' % common_exp
-    else:
-        exp_str = EXP_PRINT[print_type](common_exp)
-
-    # Possible % sign:
-    percent_str = ''
-    if '%' in options:
-        if 'L' in options:
-            # % is a special character, in LaTeX: it must be escaped.
-            #
-            # Using '\\' in the code instead of r'\' so as not to
-            # confuse emacs's syntax highlighting:
-            percent_str += ' \\'
-        percent_str += '%'
-
-    ####################
-
-    # Only true if the error should not have an exponent (has priority
-    # over common_exp):
-    special_error = not error_main or isinfinite(error_main)
-
-    # Nicer representation of the main nominal part, with no trailing
-    # zeros, when the error does not have a defined number of
-    # significant digits:
-    if special_error and fmt_parts['type'] in ('', 'g', 'G'):
-        # The main part is between 1 and 10 because any possible
-        # exponent is taken care of by common_exp, so it is
-        # formatted without an exponent (otherwise, the exponent
-        # would have to be handled for the LaTeX option):
-        fmt_suffix_n = (fmt_parts['prec'] or '')+fmt_parts['type']
-    else:
-        fmt_suffix_n = '.%d%s' % (prec, main_pres_type)
-
-
-    # print "FMT_SUFFIX_N", fmt_suffix_n
-
-    ####################
-
-    # Calculation of the mostly final numerical part value_str (no %
-    # sign, no global width applied).
-
-    # Error formatting:
-
-
-    if 'S' in options:  # Shorthand notation:
-
-        # Calculation of the uncertainty part, uncert_str:
-
-        if error_main == 0:
-            # The error is exactly zero
-            uncert_str = '0'
-        elif isnan(error_main):
-            uncert_str = robust_format(error_main, main_pres_type)
-            if 'L' in options:
-                uncert_str = r'\mathrm{%s}' % uncert_str
-        elif isinf(error_main):
-            if 'L' in options:
-                uncert_str = r'\infty'
-            else:
-                uncert_str = robust_format(error_main, main_pres_type)
-        else:  #  Error with a meaningful first digit (not 0, and real number)
-
-            uncert = round(error_main, prec)
-
-            # The representation uncert_str of the uncertainty (which will
-            # be put inside parentheses) is calculated:
-
-            # The uncertainty might straddle the decimal point: we
-            # keep it as it is, in this case (e.g. 1.2(3.4), as this
-            # makes the result easier to read); the shorthand
-            # notation then essentially coincides with the +/-
-            # notation:
-            if first_digit(uncert) >= 0 and prec > 0:
-                # This case includes a zero rounded error with digits
-                # after the decimal point:
-                uncert_str = '%.*f' % (prec, uncert)
-
-            else:
-                if uncert:
-                    # The round is important because 566.99999999 can
-                    # first be obtained when 567 is wanted (%d prints the
-                    # integer part, not the rounded value):
-                    uncert_str = '%d' % round(uncert*10.**prec)
-                else:
-                    # The decimal point indicates a truncated float
-                    # (this is easy to do, in this case, since
-                    # fmt_prefix_e is ignored):
-                    uncert_str = '0.'
-
-        # End of the final number representation (width and alignment
-        # not included). This string is important for the handling of
-        # the width:
-        value_end = '(%s)%s%s' % (uncert_str, exp_str, percent_str)
-        any_exp_factored = True  # Single exponent in the output
-
-        ##########
-        # Nominal value formatting:
-
-        # Calculation of fmt_prefix_n (prefix for the format of the
-        # main part of the nominal value):
-
-        if fmt_parts['zero'] and fmt_parts['width']:
-
-            # Padding with zeros must be done on the nominal value alone:
-
-            # Remaining width (for the nominal value):
-            nom_val_width = max(int(fmt_parts['width']) - len(value_end), 0)
-            fmt_prefix_n = '%s%s%d%s' % (
-                fmt_parts['sign'], fmt_parts['zero'], nom_val_width,
-                fmt_parts['comma'])
-
-        else:
-            # Any 'zero' part should not do anything: it is not
-            # included
-            fmt_prefix_n = fmt_parts['sign']+fmt_parts['comma']
-
-        # print "FMT_PREFIX_N", fmt_prefix_n
-        # print "FMT_SUFFIX_N", fmt_suffix_n
-
-        nom_val_str = robust_format(nom_val_main, fmt_prefix_n+fmt_suffix_n)
-
-        ##########
-        # Overriding of nom_val_str for LaTeX,; possibly based on the
-        # existing value (for NaN vs nan):
-        if 'L' in options:
-
-            if isnan(nom_val_main):
-                nom_val_str = r'\mathrm{%s}' % nom_val_str
-            elif isinf(nom_val_main):
-                # !! It is wasteful, in this case, to replace
-                # nom_val_str: could this be avoided while avoiding to
-                # duplicate the formula for nom_val_str for the common
-                # case (robust_format(...))?
-                nom_val_str = r'%s\infty' % ('-' if nom_val_main < 0 else '')
-
-        value_str = nom_val_str+value_end
-
-        # Global width, if any:
-
-        if fmt_parts['width']:  # An individual alignment is needed:
-
-            # Default alignment, for numbers: to the right (if no
-            # alignment is specified, a string is aligned to the
-            # left):
-            value_str = robust_align(
-                value_str, fmt_parts['fill'], fmt_parts['align'] or '>',
-                fmt_parts['width'])
-
-    else:  # +/- notation:
-
-        # The common exponent is factored or not, depending on the
-        # width. This gives nice columns for the nominal values and
-        # the errors (no shift due to a varying exponent), when a need
-        # is given:
-        any_exp_factored = not fmt_parts['width']
-
-        # True when the error part has any exponent directly attached
-        # (case of an individual exponent for both the nominal value
-        # and the error, when the error is a non-0, real number).
-        # The goal is to avoid the strange notation nane-10, and to
-        # avoid the 0e10 notation for an exactly zero uncertainty,
-        # because .0e can give this for a non-zero error (the goal is
-        # to have a zero uncertainty be very explicit):
-        error_has_exp = not any_exp_factored and not special_error
-
-         # Like error_has_exp, but only for real number handling
-        # (there is no special meaning to a zero nominal value):
-        nom_has_exp = not any_exp_factored and not isinfinite(nom_val_main)
-
-        # Prefix for the parts:
-        if fmt_parts['width']:  # Individual widths
-
-            # If zeros are needed, then the width is taken into
-            # account now (before the exponent is added):
-            if fmt_parts['zero']:
-
-                width = int(fmt_parts['width'])
-
-                # Remaining (minimum) width after including the
-                # exponent:
-                remaining_width = max(width-len(exp_str), 0)
-
-                fmt_prefix_n = '%s%s%d%s' % (
-                    fmt_parts['sign'], fmt_parts['zero'],
-                    remaining_width if nom_has_exp else width,
-                    fmt_parts['comma'])
-
-                fmt_prefix_e = '%s%d%s' % (
-                    fmt_parts['zero'],
-                    remaining_width if error_has_exp else width,
-                    fmt_parts['comma'])
-
-            else:
-                fmt_prefix_n = fmt_parts['sign']+fmt_parts['comma']
-                fmt_prefix_e = fmt_parts['comma']
-
-        else:  # Global width
-            fmt_prefix_n = fmt_parts['sign']+fmt_parts['comma']
-            fmt_prefix_e = fmt_parts['comma']
-
-        ## print "ANY_EXP_FACTORED", any_exp_factored
-        ## print "ERROR_HAS_EXP", error_has_exp
-        ## print "NOM_HAS_EXP", nom_has_exp
-
-        ####################
-        # Nominal value formatting:
-
-        # !! The following fails with Python < 2.6 when the format is
-        # not accepted by the % operator. This can happen when
-        # special_error is true, as the format used for the nominal
-        # value is essentially the format provided by the user, which
-        # may be empty:
-
-        # print "FMT_PREFIX_N", fmt_prefix_n
-        # print "FMT_SUFFIX_N", fmt_suffix_n
-
-        nom_val_str = robust_format(nom_val_main, fmt_prefix_n+fmt_suffix_n)
-
-        # print "NOM_VAL_STR", nom_val_str
-
-        ####################
-        # Error formatting:
-
-        # !! Note: .0f applied to a float has no decimal point, but
-        # this does not appear to be documented
-        # (http://docs.python.org/2/library/string.html#format-specification-mini-language). This
-        # feature is used anyway, because it allows a possible comma
-        # format parameter to be handled more conveniently than if the
-        # 'd' format was used.
-        #
-        # The following uses a special integer representation of a
-        # zero uncertainty:
-        if error_main:
-            # The handling of NaN/inf in the nominal value identical to
-            # the handling of NaN/inf in the standard deviation:
-            if (isinfinite(nom_val_main)
-                # Only some formats have a nicer representation:
-                and fmt_parts['type'] in ('', 'g', 'G')):
-                # The error can be formatted independently:
-                fmt_suffix_e = (fmt_parts['prec'] or '')+fmt_parts['type']
-            else:
-                fmt_suffix_e = '.%d%s' % (prec, main_pres_type)
-        else:
-            fmt_suffix_e = '.0%s' % main_pres_type
-
-        error_str = robust_format(error_main, fmt_prefix_e+fmt_suffix_e)
-
-        ##########
-        # Overriding of nom_val_str and error_str for LaTeX:
-        if 'L' in options:
-
-            if isnan(nom_val_main):
-                nom_val_str = r'\mathrm{%s}' % nom_val_str
-            elif isinf(nom_val_main):
-                nom_val_str = r'%s\infty' % ('-' if nom_val_main < 0 else '')
-
-            if isnan(error_main):
-                error_str = r'\mathrm{%s}' % error_str
-            elif isinf(error_main):
-                error_str = r'\infty'
-
-        if nom_has_exp:
-            nom_val_str += exp_str
-        if error_has_exp:
-            error_str += exp_str
-
-        ####################
-        # Final alignment of each field, if needed:
-
-        if fmt_parts['width']:  # An individual alignment is needed:
-
-            # Default alignment, for numbers: to the right (if no
-            # alignment is specified, a string is aligned to the
-            # left):
-            effective_align = fmt_parts['align'] or '>'
-
-            # robust_format() is used because it may handle alignment
-            # options, where the % operator does not:
-
-            nom_val_str = robust_align(
-                nom_val_str, fmt_parts['fill'], effective_align,
-                fmt_parts['width'])
-
-            error_str = robust_align(
-                error_str, fmt_parts['fill'], effective_align,
-                fmt_parts['width'])
-
-        ####################
-        pm_symbol = PM_SYMBOLS[print_type]  # Shortcut
-
-        ####################
-
-        # Construction of the final value, value_str, possibly with
-        # grouping (typically inside parentheses):
-
-        (LEFT_GROUPING, RIGHT_GROUPING) = GROUP_SYMBOLS[print_type]
-
-        # The nominal value and the error might have to be explicitly
-        # grouped together with parentheses, so as to prevent an
-        # ambiguous notation. This is done in parallel with the
-        # percent sign handling because this sign may too need
-        # parentheses.
-        if any_exp_factored and common_exp is not None:  # Exponent
-            value_str = ''.join((
-                LEFT_GROUPING,
-                nom_val_str, pm_symbol, error_str,
-                RIGHT_GROUPING,
-                exp_str, percent_str))
-        else:  # No exponent
-            value_str = ''.join([nom_val_str, pm_symbol, error_str])
-            if percent_str:
-                value_str = ''.join((
-                    LEFT_GROUPING, value_str, RIGHT_GROUPING, percent_str))
-            elif 'p' in options:
-                value_str = ''.join((LEFT_GROUPING, value_str, RIGHT_GROUPING))
-
-    return value_str
-
-def signif_dgt_to_limit(value, num_signif_d):
-    '''
-    Return the precision limit necessary to display value with
-    num_signif_d significant digits.
-
-    The precision limit is given as -1 for 1 digit after the decimal
-    point, 0 for integer rounding, etc. It can be positive.
-    '''
-
-    fst_digit = first_digit(value)
-
-    limit_no_rounding = fst_digit-num_signif_d+1
-
-    # The number of significant digits of the uncertainty, when
-    # rounded at this limit_no_rounding level, can be too large by 1
-    # (e.g., with num_signif_d = 1, 0.99 gives limit_no_rounding = -1, but
-    # the rounded value at that limit is 1.0, i.e. has 2
-    # significant digits instead of num_signif_d = 1). We correct for
-    # this effect by adjusting limit if necessary:
-    rounded = round(value, -limit_no_rounding)
-    fst_digit_rounded = first_digit(rounded)
-
-    if fst_digit_rounded > fst_digit:
-        # The rounded limit is fst_digit_rounded-num_signif_d+1;
-        # but this can only be 1 above the non-rounded limit:
-        limit_no_rounding += 1
-
-    return limit_no_rounding
 
 class LinearCombination(object):
     """
@@ -2827,323 +1885,6 @@ class Variable(AffineScalarFunc):
         # Reference: http://www.doughellmann.com/PyMOTW/copy/index.html
 
         return self.__copy__()
-
-
-###############################################################################
-
-# Utilities
-
-def nominal_value(x):
-    """
-    Return the nominal value of x if it is a quantity with
-    uncertainty (i.e., an AffineScalarFunc object); otherwise, returns
-    x unchanged.
-
-    This utility function is useful for transforming a series of
-    numbers, when only some of them generally carry an uncertainty.
-    """
-
-    if isinstance(x, AffineScalarFunc):
-        return x.nominal_value
-    else:
-        return x
-
-def std_dev(x):
-    """
-    Return the standard deviation of x if it is a quantity with
-    uncertainty (i.e., an AffineScalarFunc object); otherwise, returns
-    the float 0.
-
-    This utility function is useful for transforming a series of
-    numbers, when only some of them generally carry an uncertainty.
-    """
-
-    if isinstance(x, AffineScalarFunc):
-        return x.std_dev
-    else:
-        return 0.
-
-def covariance_matrix(nums_with_uncert):
-    """
-    Return a matrix that contains the covariances between the given
-    sequence of numbers with uncertainties (AffineScalarFunc objects).
-    The resulting matrix implicitly depends on their ordering in
-    'nums_with_uncert'.
-
-    The covariances are floats (never int objects).
-
-    The returned covariance matrix is the exact linear approximation
-    result, if the nominal values of the numbers with uncertainties
-    and of their variables are their mean.  Otherwise, the returned
-    covariance matrix should be close to its linear approximation
-    value.
-
-    The returned matrix is a list of lists.
-    """
-    # See PSI.411 in EOL's notes.
-
-    covariance_matrix = []
-    for (i1, expr1) in enumerate(nums_with_uncert, 1):
-        derivatives1 = expr1.derivatives  # Optimization
-        vars1 = set(derivatives1)  # !! Python 2.7+: viewkeys() would work
-        coefs_expr1 = []
-
-        for expr2 in nums_with_uncert[:i1]:
-            derivatives2 = expr2.derivatives  # Optimization
-            coefs_expr1.append(sum(
-                ((derivatives1[var]*derivatives2[var]*var._std_dev**2)
-                # var is a variable common to both numbers with
-                # uncertainties:
-                for var in vars1.intersection(derivatives2)),
-                # The result is always a float (sum() with no terms
-                # returns an integer):
-                0.))
-
-        covariance_matrix.append(coefs_expr1)
-
-    # We symmetrize the matrix:
-    for (i, covariance_coefs) in enumerate(covariance_matrix):
-        covariance_coefs.extend([covariance_matrix[j][i]
-                                 for j in range(i+1, len(covariance_matrix))])
-
-    return covariance_matrix
-
-try:
-    import numpy
-except ImportError:
-    pass
-else:
-    def correlation_matrix(nums_with_uncert):
-        '''
-        Return the correlation matrix of the given sequence of
-        numbers with uncertainties, as a NumPy array of floats.
-        '''
-
-        cov_mat = numpy.array(covariance_matrix(nums_with_uncert))
-
-        std_devs = numpy.sqrt(cov_mat.diagonal())
-
-        return cov_mat/std_devs/std_devs[numpy.newaxis].T
-
-    __all__.append('correlation_matrix')
-
-###############################################################################
-# Parsing of values with uncertainties:
-
-# Parsing of (part of) numbers. The reason why the decimal part is
-# parsed (if any), instead of using the parsing built in float(), is
-# that the presence (or not) of a decimal point does matter, in the
-# semantics of some representations (e.g. .1(2.) = .1+/-2, whereas
-# .1(2) = .1+/-0.2), so just getting the numerical value of the part
-# in parentheses would not be sufficient.
-POSITIVE_DECIMAL_UNSIGNED_OR_NON_FINITE = r'((\d*)(\.\d*)?|nan|NAN|inf|INF)'
-
-# Regexp for a number with uncertainty (e.g., "-1.234(2)e-6"), where
-# the uncertainty is optional (in which case the uncertainty is
-# implicit). The uncertainty can also be nan or NAN:
-#
-# !! WARNING: in Python 2, the code relies on "… % <unicode string>" returning
-# a Unicode string (even if the template is not Unicode):
-NUMBER_WITH_UNCERT_RE_STR = u'''
-    ([+-])?  # Sign
-    %s  # Main number
-    (?:\\(%s\\))?  # Optional uncertainty
-    (?:
-        (?:[eE]|\\s*×\\s*10)
-        (.*)
-    )?  # Optional exponent
-    ''' % (POSITIVE_DECIMAL_UNSIGNED_OR_NON_FINITE,
-           POSITIVE_DECIMAL_UNSIGNED_OR_NON_FINITE)
-
-NUMBER_WITH_UNCERT_RE_MATCH = re.compile(
-    u"%s$" % NUMBER_WITH_UNCERT_RE_STR, re.VERBOSE).match
-
-# Number with uncertainty with a factored exponent (e.g., of the form
-# (... +/- ...)e10): this is a loose matching, so as to accommodate
-# for multiple formats:
-NUMBER_WITH_UNCERT_GLOBAL_EXP_RE_MATCH = re.compile(u'''
-    \\(
-    (?P<simple_num_with_uncert>.*)
-    \\)
-    (?:[eE]|\\s*×\\s*10) (?P<exp_value>.*)
-    $''', re.VERBOSE).match
-
-class NotParenUncert(ValueError):
-    '''
-    Raised when a string representing an exact number or a number with
-    an uncertainty indicated between parentheses was expected but not
-    found.
-    '''
-
-def parse_error_in_parentheses(representation):
-    # !!!! The code seems to handle superscript exponents, but the
-    # docstring doesn't reflect this!?
-    """
-    Return (value, error) from a string representing a number with
-    uncertainty like 12.34(5), 12.34(142), 12.5(3.4), 12.3(4.2)e3, or
-    13.4(nan)e10.  If no parenthesis is given, an uncertainty of one
-    on the last digit is assumed.
-
-    The digits between parentheses correspond to the same number of digits
-    at the end of the nominal value (the decimal point in the uncertainty
-    is optional). Example: 12.34(142) = 12.34±1.42.
-    
-    Raises ValueError if the string cannot be parsed.
-    """
-
-    match = NUMBER_WITH_UNCERT_RE_MATCH(representation)
-
-    if match:
-        # The 'main' part is the nominal value, with 'int'eger part, and
-        # 'dec'imal part.  The 'uncert'ainty is similarly broken into its
-        # integer and decimal parts.
-        (sign, main, _, main_dec, uncert, uncert_int, uncert_dec,
-         exponent) = match.groups()
-    else:
-        raise NotParenUncert("Unparsable number representation: '%s'."
-                             " See the documentation of ufloat_fromstr()."
-                             % representation)
-
-    # Global exponent:
-    if exponent:
-        factor = 10.**nrmlze_superscript(exponent)
-    else:
-        factor = 1
-
-    # Nominal value:
-    value = float((sign or '')+main)*factor
-
-    if uncert is None:
-        # No uncertainty was found: an uncertainty of 1 on the last
-        # digit is assumed:
-        uncert_int = '1'  # The other parts of the uncertainty are None
-
-    # Do we have a fully explicit uncertainty?
-    if uncert_dec is not None or uncert in {'nan', 'NAN', 'inf', 'INF'}:
-        uncert_value = float(uncert)
-    else:
-        # uncert_int represents an uncertainty on the last digits:
-
-        # The number of digits after the period defines the power of
-        # 10 that must be applied to the provided uncertainty:
-        if main_dec is None:
-            num_digits_after_period = 0
-        else:
-            num_digits_after_period = len(main_dec)-1
-
-        uncert_value = int(uncert_int)/10.**num_digits_after_period
-
-    # We apply the exponent to the uncertainty as well:
-    uncert_value *= factor
-
-    return (value, uncert_value)
-
-# Regexp for catching the two variable parts of -1.2×10⁻¹²:
-PRETTY_PRINT_MATCH = re.compile(u'(.*?)\\s*×\\s*10(.*)').match
-
-def to_float(value_str):
-    '''
-    Converts a string representing a float to a float.
-
-    The usual valid Python float() representations are correctly
-    parsed.
-
-    In addition, the pretty-print notation -1.2×10⁻¹² is also
-    converted.
-
-    ValueError is raised if no float can be obtained.
-    '''
-
-    try:
-        return float(value_str)
-    except ValueError:
-        pass
-
-    # The pretty-print notation is tried:
-    match = PRETTY_PRINT_MATCH(value_str)
-    if match:
-        try:
-            return float(match.group(1))*10.**nrmlze_superscript(match.group(2))
-        except ValueError:
-            raise ValueError('Mantissa or exponent incorrect in pretty-print'
-                             ' form %s' % value_str)
-    else:
-        raise ValueError('No valid Python float or pretty-print form'
-                         ' recognized in %s' % value_str)
-
-
-cannot_parse_ufloat_msg_pat = (
-    'Cannot parse %s: see the documentation for ufloat_fromstr() for a'
-    ' list of accepted formats')
-
-# The following function is not exposed because it can in effect be
-# obtained by doing x = ufloat_fromstr(representation) and reading
-# x.nominal_value and x.std_dev:
-def str_to_number_with_uncert(representation):
-    """
-    Given a string that represents a number with uncertainty, returns the
-    nominal value and the uncertainty.
-
-    See the documentation for ufloat_fromstr() for a list of accepted
-    formats.
-
-    When no numerical error is given, an uncertainty of 1 on the last
-    digit is implied.
-
-    Raises ValueError if the string cannot be parsed.
-
-    representation -- string with no leading or trailing spaces.
-    """
-
-    # The "p" format can add parentheses around the whole printed result: we
-    # remove them:
-    if representation.startswith('(') and representation.endswith(')'):
-        representation = representation[1:-1]
-
-    match = NUMBER_WITH_UNCERT_GLOBAL_EXP_RE_MATCH(representation)
-
-    # The representation is simplified, but the global factor is
-    # calculated:
-
-    if match:
-
-        # We have a form with a factored exponent: (1.23 +/- 0.01)e10,
-        # etc.
-
-        exp_value_str = match.group('exp_value')
-
-        try:
-            exponent = nrmlze_superscript(exp_value_str)
-        except ValueError:
-            raise ValueError(cannot_parse_ufloat_msg_pat % representation)
-
-        factor = 10.**exponent
-
-        representation = match.group('simple_num_with_uncert')
-    else:
-        factor = 1  # No global exponential factor
-
-    match = re.match(u'(.*)(?:\\+/-|±)(.*)', representation)
-    if match:
-
-        (nom_value, uncert) = match.groups()
-
-        try:
-            # Simple form 1234.45+/-1.2 or 1234.45±1.2, or 1.23e-10+/-1e-23
-            # or -1.2×10⁻¹²±1e23:
-            parsed_value = (to_float(nom_value)*factor,
-                            to_float(uncert)*factor)
-        except ValueError:
-            raise ValueError(cannot_parse_ufloat_msg_pat % representation)
-
-    else:
-        # Form with error parentheses or no uncertainty:
-        try:
-            parsed_value = parse_error_in_parentheses(representation)
-        except NotParenUncert:
-            raise ValueError(cannot_parse_ufloat_msg_pat % representation)
-
-    return parsed_value
 
 def ufloat_fromstr(representation, tag=None):
     """
